genero el submit binario porque sino es imposible subir a Kaggle (el archivo binario se puede comprimir a 11Mb y el otro con probabilidades pesa 300MB comprimido)

+++ san-francisco-crime-classification/src/CsvWriter.cpp
@@ -12,7 +12,7 @@ CsvWriter::CsvWriter() {
 
 }
 
-void CsvWriter::makeSubmitWithMatrix(string path,MatrixXf matrix){
+void CsvWriter::makeSubmitWithMatrix(string path,MatrixXi matrix){
 
 	ofstream file(path);
 	int rows = matrix.rows();
